Большой плеер и его функционал

+++ src/app/tools/soundcloud/soundcloud.service.ts
@@ -11,9 +11,11 @@ export class SoundcloudService {
 
 
   public onFinished: EventEmitter<Track>;
+  public onTimeChanged: EventEmitter<number>;
 
   constructor(private http: Http) {
     this.onFinished = new EventEmitter<Track>();
+    this.onTimeChanged = new EventEmitter<number>();
 
     SC.initialize({
       client_id: ConfigurationProvider.soundcloudKey
@@ -34,23 +36,41 @@ export class SoundcloudService {
   private player;
   public playingTrack: Track;
 
-  public play(track: Track){
+  public play(track: Track, volume: number){
+
+    if(this.playingTrack && track.id == this.playingTrack.id && this.player && this.player.getState() == "paused"){
+      this.player.play();
+      return;
+    }
+
     SC.stream(track.url.replace("https://api.soundcloud.com/", "")).then((player)=>{
+      if(this.player) this.player.kill();
       this.player = player;
-      //player.play();
+      this.player.setVolume(volume/100);
+      player.play();
       var that = this;
       this.playingTrack = track;
       player.on("finish", ()=> that.onFinished.emit(track));
+      player.on("time", (event)=> that.onTimeChanged.emit(event));
     });
   }
 
   public stop(){
     if(this.player){
-      this.player.kill();
-      this.playingTrack = undefined;
+      this.player.pause();
     }
   }
 
+  public updateVolume(volume: number){
+    if(!this.player) return;
 
+    this.player.setVolume(volume/100);
+  }
+
+  public setTime(timeMs: number){
+    if(this.player){
+      this.player.seek(timeMs);
+    }
+  }
 
 }
